changes after code review
