+++ adapter/pkg/oasparser/configGenerator.go
@@ -40,14 +40,13 @@ func GetProductionRoutesClustersEndpoints(byteArr []byte, upstreamCerts []byte, 
 
 	if apiType == mgw.HTTP {
 		mgwSwagger = operator.GetMgwSwagger(byteArr)
-		routes, clusters, endpoints = envoy.CreateRoutesWithClusters(mgwSwagger, upstreamCerts)
 	} else if apiType == mgw.WS {
 		mgwSwagger = operator.GetMgwSwaggerWebSocket(byteArr)
-		routes, clusters, endpoints = envoy.CreateRouteWithClustersWebSocket(mgwSwagger, upstreamCerts)
 	} else {
 		// Unreachable else condition. Added in case previous apiType check fails due to any modifications.
 		loggers.LoggerOasparser.Errorf("API type not currently supported with WSO2 Micro-gateway")
 	}
+	routes, clusters, endpoints = envoy.CreateRoutesWithClusters(mgwSwagger, upstreamCerts)
 	//TODO: (VirajSalaka) Decide if this needs to be added to the MgwSwagger
 
 	return routes, clusters, endpoints
