+++ adapter/pkg/oasparser/envoyconf/envoyconf_internal_test.go
@@ -73,6 +73,7 @@ func TestCreateRoute(t *testing.T) {
 	xWso2BasePath := "/xWso2BasePath"
 	basePath := "/basepath"
 	title := "WSO2"
+	apiType := "HTTP"
 	endpoint := model.Endpoint{
 		Host:     "abc.com",
 		Basepath: basePath,
@@ -126,7 +127,7 @@ func TestCreateRoute(t *testing.T) {
 		},
 	}
 
-	generatedRouteWithXWso2BasePath := createRoute(title, xWso2BasePath, version, endpoint.Basepath, resourceWithGet, clusterName, "")
+	generatedRouteWithXWso2BasePath := createRoute(title, apiType, xWso2BasePath, version, endpoint.Basepath, resourceWithGet.GetPath(), resourceWithGet.GetMethod(), clusterName, "")
 	assert.NotNil(t, generatedRouteWithXWso2BasePath, "Route should not be null.")
 	assert.Equal(t, expectedRouteActionWithXWso2BasePath, generatedRouteWithXWso2BasePath.Action,
 		"Route generation mismatch when xWso2BasePath option is provided.")
@@ -134,7 +135,7 @@ func TestCreateRoute(t *testing.T) {
 	assert.Equal(t, "^(GET)$", generatedRouteWithXWso2BasePath.GetMatch().Headers[0].GetSafeRegexMatch().Regex,
 		"Assigned HTTP Method Regex is incorrect when single method is available.")
 
-	generatedRouteWithoutXWso2BasePath := createRoute(title, "", version, endpoint.Basepath, resourceWithGetPost, clusterName, "")
+	generatedRouteWithoutXWso2BasePath := createRoute(title, apiType, "", version, endpoint.Basepath, resourceWithGetPost.GetPath(), resourceWithGetPost.GetMethod(), clusterName, "")
 	assert.NotNil(t, generatedRouteWithoutXWso2BasePath, "Route should not be null")
 	assert.NotNil(t, generatedRouteWithoutXWso2BasePath.GetMatch().Headers, "Headers property should not be null")
 	assert.Equal(t, "^(GET|POST)$", generatedRouteWithoutXWso2BasePath.GetMatch().Headers[0].GetSafeRegexMatch().Regex,
@@ -157,23 +158,24 @@ func TestCreateRouteClusterSpecifier(t *testing.T) {
 	endpointBasePath := "/basepath"
 	title := "WSO2"
 	version := "1.0.0"
+	apiType := "HTTP"
 
 	resourceWithGet := model.CreateMinimalDummyResourceForTests("/resourcePath", []string{"GET"},
 		"resource_operation_id", []model.Endpoint{}, []model.Endpoint{})
 
-	routeWithProdEp := createRoute(title, xWso2BasePath, version, endpointBasePath, resourceWithGet, prodClusterName, "")
+	routeWithProdEp := createRoute(title, apiType, xWso2BasePath, version, endpointBasePath, resourceWithGet.GetPath(), resourceWithGet.GetMethod(), prodClusterName, "")
 	assert.NotNil(t, routeWithProdEp, "Route should not be null")
 	assert.NotNil(t, routeWithProdEp.GetRoute().GetCluster(), "Route Cluster Name should not be null.")
 	assert.Empty(t, routeWithProdEp.GetRoute().GetClusterHeader(), "Route Cluster Header should be empty.")
 	assert.Equal(t, prodClusterName, routeWithProdEp.GetRoute().GetCluster(), "Route Cluster Name mismatch.")
 
-	routeWithSandEp := createRoute(title, xWso2BasePath, version, endpointBasePath, resourceWithGet, "", sandClusterName)
+	routeWithSandEp := createRoute(title, apiType, xWso2BasePath, version, endpointBasePath, resourceWithGet.GetPath(), resourceWithGet.GetMethod(), "", sandClusterName)
 	assert.NotNil(t, routeWithSandEp, "Route should not be null")
 	assert.NotNil(t, routeWithSandEp.GetRoute().GetCluster(), "Route Cluster Name should not be null.")
 	assert.Empty(t, routeWithSandEp.GetRoute().GetClusterHeader(), "Route Cluster Header should be empty.")
 	assert.Equal(t, sandClusterName, routeWithSandEp.GetRoute().GetCluster(), "Route Cluster Name mismatch.")
 
-	routeWithProdSandEp := createRoute(title, xWso2BasePath, version, endpointBasePath, resourceWithGet, prodClusterName,
+	routeWithProdSandEp := createRoute(title, apiType, xWso2BasePath, version, endpointBasePath, resourceWithGet.GetPath(), resourceWithGet.GetMethod(), prodClusterName,
 		sandClusterName)
 	assert.NotNil(t, routeWithProdSandEp, "Route should not be null")
 	assert.NotNil(t, routeWithProdSandEp.GetRoute().GetClusterHeader(), "Route Cluster Header should not be null.")
@@ -192,11 +194,12 @@ func TestCreateRouteExtAuthzContext(t *testing.T) {
 	endpointBasePath := "/basepath"
 	title := "WSO2"
 	version := "1.0.0"
+	apiType := "HTTP"
 
 	resourceWithGet := model.CreateMinimalDummyResourceForTests("/resourcePath", []string{"GET"},
 		"resource_operation_id", []model.Endpoint{}, []model.Endpoint{})
 
-	routeWithProdEp := createRoute(title, xWso2BasePath, version, endpointBasePath, resourceWithGet, prodClusterName, sandClusterName)
+	routeWithProdEp := createRoute(title, apiType, xWso2BasePath, version, endpointBasePath, resourceWithGet.GetPath(), resourceWithGet.GetMethod(), prodClusterName, sandClusterName)
 	assert.NotNil(t, routeWithProdEp, "Route should not be null")
 	assert.NotNil(t, routeWithProdEp.GetTypedPerFilterConfig(), "TypedPerFilter config should not be null")
 	assert.NotNil(t, routeWithProdEp.GetTypedPerFilterConfig()[wellknown.HTTPExternalAuthorization],
