+++ adapter/pkg/oasparser/model/openApi_internal_test.go
@@ -44,10 +44,10 @@ func TestSetInfoOpenAPI(t *testing.T) {
 			},
 
 			MgwSwagger{
-				swaggerVersion: "openApi",
-				description:    "Swagger definition",
-				title:          "petsore",
-				version:        "1.0",
+				apiType:     "HTTP",
+				description: "Swagger definition",
+				title:       "petsore",
+				version:     "1.0",
 			},
 			"usual case",
 		},
@@ -57,11 +57,11 @@ func TestSetInfoOpenAPI(t *testing.T) {
 				Info:    nil,
 			},
 			MgwSwagger{
-				id:             "",
-				swaggerVersion: "openApi",
-				description:    "",
-				title:          "",
-				version:        "",
+				id:          "",
+				apiType:     "HTTP",
+				description: "",
+				title:       "",
+				version:     "",
 			},
 			"when info section is null",
 		},
