+++ adapter/pkg/oasparser/model/swagger_internal_test.go
@@ -48,11 +48,11 @@ func TestSetInfoSwagger(t *testing.T) {
 			},
 
 			MgwSwagger{
-				id:             "v1",
-				swaggerVersion: "swagger",
-				description:    "Swagger definition",
-				title:          "petsore",
-				version:        "1.0",
+				id:          "v1",
+				apiType:     "HTTP",
+				description: "Swagger definition",
+				title:       "petsore",
+				version:     "1.0",
 			},
 			"usual case",
 		},
@@ -66,11 +66,11 @@ func TestSetInfoSwagger(t *testing.T) {
 				},
 			},
 			MgwSwagger{
-				id:             "",
-				swaggerVersion: "swagger",
-				description:    "",
-				title:          "",
-				version:        "",
+				id:          "",
+				apiType:     "HTTP",
+				description: "",
+				title:       "",
+				version:     "",
 			},
 			"when info section is null",
 		},
